backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "pandas>=3.0.1",
    "pdfplumber>=0.11.9",
    "pikepdf>=10.5.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "supabase>=2.28.2",
    "uvicorn>=0.41.0",
]
